Add __version__

+++ setup.py
@@ -2,6 +2,16 @@
 from setuptools import setup
 import os
 
+# get the version (don't import, so dependencies are not needed)
+version = None
+with open('microstates.py') as fid:
+    for line in (line.strip() for line in fid):
+        if line.startswith('__version__'):
+            version = line.split('=')[1].strip().strip('\'')
+            break
+if version is None:
+    raise RuntimeError('Could not determine version')
+
 if __name__ == "__main__":
     if os.path.exists('MANIFEST'):
         os.remove('MANIFEST')
@@ -10,10 +20,11 @@ if __name__ == "__main__":
         name='mne_microstates',
         maintainer='Marijn van Vliet',
         maintainer_email='w.m.vanvliet@gmail.com',
-        description='Code for microstate analysis, in combination with MNE-Python.',
+        description=('Code for microstate analysis, in combination with '
+                     'MNE-Python.'),
         license='BSD-3',
         url='https://github.com/wmvanvliet/mne_microstates',
-        version='0.2',
+        version=version,
         long_description=open('README.md').read(),
         classifiers=['Intended Audience :: Science/Research',
                      'Intended Audience :: Developers',
